Crate rye

A custom unit testing framework for Rust.

The concept is heavily influenced by the section mechanism in Catch2, a C++ unit testing framework library.

Macros

fail
Mark the current test case as failed and then terminate its execution.

skip
Mark the current test case as skipped and then terminate its execution.

test_harness
Generate the main function for running test application.

Structs

Context
Context values while running the test case.

Session

Traits

Termination
Arbitrary return values in test cases.

Attribute Macros

test
Generate a single test case.

test_main
Define a test main function.